refactor(app): share interval feature traversal
